WIP re-architect how we initialize TUN
