+++ rust/connlib/clients/android/Cargo.toml
@@ -26,6 +26,7 @@ serde_json = "1"
 thiserror = "1"
 url = "2.4.0"
 tokio = { version = "1.36", default-features = false, features = ["rt"] }
+firezone-tunnel = { workspace = true }
 
 [target.'cfg(target_os = "android")'.dependencies]
 tracing-android = "0.2"
