+++ rust/connlib/clients/apple/Cargo.toml
@@ -25,6 +25,7 @@ tracing-subscriber = "0.3"
 tracing-appender = "0.2"
 url = "2.5.0"
 tokio = { version = "1.36", default-features = false, features = ["rt"] }
+firezone-tunnel = { workspace = true }
 
 [lib]
 name = "connlib"
