+++ rust/connlib/tunnel/src/lib.rs
@@ -16,6 +16,7 @@ use std::{
 };
 
 pub use client::{ClientState, Request};
+pub use device_channel::Tun;
 pub use gateway::{GatewayState, ResolvedResourceDescriptionDns};
 
 mod client;
