+++ rust/linux-client/src/main.rs
@@ -7,9 +7,25 @@ use connlib_shared::{
     LoginUrl,
 };
 use firezone_cli_utils::{setup_global_subscriber, CommonArgs};
+use firezone_tunnel::Tun;
+use netlink_packet_route::route::{RouteProtocol, RouteScope};
+use netlink_packet_route::rule::RuleAction;
+use rtnetlink::{new_connection, Error::NetlinkError, Handle};
+use rtnetlink::{RouteAddRequest, RuleAddRequest};
 use secrecy::SecretString;
-use std::{future, net::IpAddr, path::PathBuf, str::FromStr, task::Poll};
+use std::{
+    future,
+    net::{IpAddr, Ipv4Addr, Ipv6Addr},
+    path::PathBuf,
+    str::FromStr,
+    task::Poll,
+};
 use tokio::signal::unix::SignalKind;
+
+const IFACE_NAME: &str = "tun-firezone";
+const DEFAULT_MTU: u32 = 1280;
+const FILE_ALREADY_EXISTS: i32 = -17;
+const FIREZONE_TABLE: u32 = 0x2021_fd00;
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -19,10 +35,17 @@ async fn main() -> Result<()> {
     let (layer, handle) = cli.log_dir.as_deref().map(file_logger::layer).unzip();
     setup_global_subscriber(layer);
 
+    let (connection, netlink_handle, _) = new_connection()?;
+    tokio::spawn(connection);
+
+    let (new_tun_sender, mut new_tun_receiver) = tokio::sync::mpsc::channel(1);
+
     let dns_control_method = get_dns_control_from_env();
     let callbacks = CallbackHandler {
         dns_control_method: dns_control_method.clone(),
         handle,
+        new_tun_sender,
+        netlink_handle,
     };
 
     // AKA "Device ID", not the Firezone slug
@@ -67,6 +90,11 @@ async fn main() -> Result<()> {
             continue;
         }
 
+        if let Poll::Ready(Some(tun)) = new_tun_receiver.poll_recv(cx) {
+            session.update_tun(tun);
+            continue;
+        }
+
         return Poll::Pending;
     })
     .await;
@@ -80,6 +108,9 @@ async fn main() -> Result<()> {
 struct CallbackHandler {
     dns_control_method: Option<DnsControlMethod>,
     handle: Option<file_logger::Handle>,
+    new_tun_sender: tokio::sync::mpsc::Sender<Tun>,
+
+    netlink_handle: Handle,
 }
 
 impl Callbacks for CallbackHandler {
@@ -121,6 +152,245 @@ impl Callbacks for CallbackHandler {
                 tracing::debug!("Failed to roll over to new file: {e}");
                 None
             })
+    }
+
+    fn on_set_interface_config(
+        &self,
+        ip4: std::net::Ipv4Addr,
+        ip6: std::net::Ipv6Addr,
+        dns_servers: Vec<IpAddr>,
+    ) {
+        tokio::spawn(set_iface_config(
+            ip4,
+            ip6,
+            dns_servers,
+            self.netlink_handle.clone(),
+            self.dns_control_method,
+        ));
+    }
+
+    fn on_update_routes(&self, _: Vec<connlib_shared::Cidrv4>, _: Vec<connlib_shared::Cidrv6>) {
+        tokio::spawn(set_routes(self.netlink_handle.clone()));
+    }
+}
+
+async fn set_routes(handle: Handle) -> Result<()> {
+    let index = handle
+        .link()
+        .get()
+        .match_name(IFACE_NAME.to_string())
+        .execute()
+        .try_next()
+        .await?
+        .ok_or(Error::NoIface)?
+        .header
+        .index;
+
+    for route in new_routes.difference(&current_routes) {
+        add_route(route, index, &handle).await;
+    }
+
+    for route in current_routes.difference(&new_routes) {
+        delete_route(route, index, &handle).await;
+    }
+
+    Ok(())
+}
+
+#[tracing::instrument(level = "trace", skip(handle))]
+async fn set_iface_config(
+    ipv4: std::net::Ipv4Addr,
+    ipv6: std::net::Ipv6Addr,
+    dns_config: Vec<IpAddr>,
+    handle: Handle,
+    dns_control_method: Option<DnsControlMethod>,
+) -> Result<()> {
+    let index = handle
+        .link()
+        .get()
+        .match_name(IFACE_NAME.to_string())
+        .execute()
+        .try_next()
+        .await?
+        .ok_or(Error::NoIface)?
+        .header
+        .index;
+
+    let ips = handle
+        .address()
+        .get()
+        .set_link_index_filter(index)
+        .execute();
+
+    ips.try_for_each(|ip| handle.address().del(ip).execute())
+        .await?;
+
+    handle.link().set(index).mtu(1280).execute().await?;
+
+    let res_v4 = handle.address().add(index, ipv4.into(), 32).execute().await;
+    let res_v6 = handle
+        .address()
+        .add(index, ipv6.into(), 128)
+        .execute()
+        .await;
+
+    handle.link().set(index).up().execute().await?;
+
+    if res_v4.is_ok() {
+        if let Err(e) = make_rule(&handle).v4().execute().await {
+            if !matches!(&e, NetlinkError(err) if err.raw_code() == FILE_ALREADY_EXISTS) {
+                tracing::warn!(
+                    "Couldn't add ip rule for ipv4: {e:?}, ipv4 packets won't be routed"
+                );
+            }
+            // TODO: Be smarter about this
+        } else {
+            tracing::debug!("Successfully created ip rule for ipv4");
+        }
+    }
+
+    if res_v6.is_ok() {
+        if let Err(e) = make_rule(&handle).v6().execute().await {
+            if !matches!(&e, NetlinkError(err) if err.raw_code() == FILE_ALREADY_EXISTS) {
+                tracing::warn!(
+                    "Couldn't add ip rule for ipv6: {e:?}, ipv6 packets won't be routed"
+                );
+            }
+            // TODO: Be smarter about this
+        } else {
+            tracing::debug!("Successfully created ip rule for ipv6");
+        }
+    }
+
+    res_v4.or(res_v6)?;
+
+    match dns_control_method {
+        None => {}
+        Some(DnsControlMethod::EtcResolvConf) => {
+            etc_resolv_conf::configure_dns(&dns_config).await?
+        }
+        Some(DnsControlMethod::NetworkManager) => configure_network_manager(&dns_config).await?,
+        Some(DnsControlMethod::Systemd) => configure_systemd_resolved(&dns_config).await?,
+    }
+
+    // TODO: Having this inside the library is definitely wrong. I think `set_iface_config`
+    // needs to return before `new` returns, so that the `on_tunnel_ready` callback
+    // happens after the IP address and DNS are set up. Then we can call `sd_notify`
+    // inside `on_tunnel_ready` in the client.
+    //
+    // `sd_notify::notify` is always safe to call, it silently returns `Ok(())`
+    // if we aren't running as a systemd service.
+    if let Err(error) = sd_notify::notify(true, &[sd_notify::NotifyState::Ready]) {
+        // Nothing we can do about it
+        tracing::warn!(?error, "Failed to notify systemd that we're ready");
+    }
+
+    Ok(())
+}
+
+async fn configure_network_manager(_dns_config: &[IpAddr]) -> Result<()> {
+    Err(Error::Other(
+        "DNS control with NetworkManager is not implemented yet",
+    ))
+}
+
+async fn configure_systemd_resolved(dns_config: &[IpAddr]) -> Result<()> {
+    let status = tokio::process::Command::new("resolvectl")
+        .arg("dns")
+        .arg(IFACE_NAME)
+        .args(dns_config.iter().map(ToString::to_string))
+        .status()
+        .await
+        .map_err(|_| Error::ResolvectlFailed)?;
+    if !status.success() {
+        return Err(Error::ResolvectlFailed);
+    }
+
+    let status = tokio::process::Command::new("resolvectl")
+        .arg("domain")
+        .arg(IFACE_NAME)
+        .arg("~.")
+        .status()
+        .await
+        .map_err(|_| Error::ResolvectlFailed)?;
+    if !status.success() {
+        return Err(Error::ResolvectlFailed);
+    }
+
+    tracing::info!(?dns_config, "Configured DNS sentinels with `resolvectl`");
+
+    Ok(())
+}
+
+fn make_rule(handle: &Handle) -> RuleAddRequest {
+    let mut rule = handle
+        .rule()
+        .add()
+        .fw_mark(FIREZONE_MARK)
+        .table_id(FIREZONE_TABLE)
+        .action(RuleAction::ToTable);
+
+    rule.message_mut()
+        .header
+        .flags
+        .push(netlink_packet_route::rule::RuleFlag::Invert);
+
+    rule.message_mut()
+        .attributes
+        .push(netlink_packet_route::rule::RuleAttribute::Protocol(
+            RouteProtocol::Kernel,
+        ));
+
+    rule
+}
+
+fn make_route(idx: u32, handle: &Handle) -> RouteAddRequest {
+    handle
+        .route()
+        .add()
+        .output_interface(idx)
+        .protocol(RouteProtocol::Static)
+        .scope(RouteScope::Universe)
+        .table_id(FIREZONE_TABLE)
+}
+
+fn make_route_v4(idx: u32, handle: &Handle, route: Ipv4Network) -> RouteAddRequest<Ipv4Addr> {
+    make_route(idx, handle)
+        .v4()
+        .destination_prefix(route.network_address(), route.netmask())
+}
+
+fn make_route_v6(idx: u32, handle: &Handle, route: Ipv6Network) -> RouteAddRequest<Ipv6Addr> {
+    make_route(idx, handle)
+        .v6()
+        .destination_prefix(route.network_address(), route.netmask())
+}
+
+async fn add_route(route: &IpNetwork, idx: u32, handle: &Handle) {
+    let res = match route {
+        IpNetwork::V4(ipnet) => make_route_v4(idx, handle, *ipnet).execute().await,
+        IpNetwork::V6(ipnet) => make_route_v6(idx, handle, *ipnet).execute().await,
+    };
+
+    match res {
+        Ok(_) => {}
+        Err(NetlinkError(err)) if err.raw_code() == FILE_ALREADY_EXISTS => {}
+        // TODO: we should be able to surface this error and handle it depending on
+        // if any of the added routes succeeded.
+        Err(err) => {
+            tracing::error!(%route, "failed to add route: {err}");
+        }
+    }
+}
+
+async fn delete_route(route: &IpNetwork, idx: u32, handle: &Handle) {
+    let message = match route {
+        IpNetwork::V4(ipnet) => make_route_v4(idx, handle, *ipnet).message_mut().clone(),
+        IpNetwork::V6(ipnet) => make_route_v6(idx, handle, *ipnet).message_mut().clone(),
+    };
+
+    if let Err(err) = handle.route().del(message).execute().await {
+        tracing::error!(%route, "failed to add route: {err:#?}");
     }
 }
